cloning: hw/9pfs/9p.c: Fix fsync requests when using WSTAT for 9p2000.u version
When calling fsync() on a Linux VM for 9pfs protocol version 9p2000.u, the
backend sends back to the frontend an 9p response with an invalid header
containing the size field equal to zero. The response is discarded by the
guest, given that a valid response header should at least contain the 9p
header size. Because of this, the guest blocks waiting for a valid response
from the backend. This patch fixes this behavior so that the backend will
send a valid RWSTAT response for fsync() calls.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>